feat(config): move AppConfig and WorkerConfig to config crate (#61)
