+++ crates/config/src/lib.rs
@@ -1,9 +1,13 @@
+pub mod app;
 pub mod database;
 pub mod error;
 pub mod helpers;
 pub mod s3;
+pub mod worker;
 
+pub use app::AppConfig;
 pub use database::DatabaseConfig;
 pub use error::ConfigError;
 pub use helpers::{optional_env, optional_env_or, parse_env_or, required_env};
 pub use s3::S3Config;
+pub use worker::WorkerConfig;
